Migrate env cmds to sys module
